backend/requirements.txt
# FastAPI e servidor
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Banco de dados
sqlalchemy==2.0.36
alembic==1.17.2
psycopg2-binary==2.9.10

# Utilitários
python-dotenv==1.0.1
pydantic==2.10.4

# Validação e tipos
typing-extensions==4.12.2
